refactor(backend): unify entry points to server.ts (CD-1)
- Move CORS allowedOrigins from index.ts to server.ts
- Update vercel.json rewrites from /api/index.js to /api/server.js
- Delete src/index.ts (dead, diverged from server.ts)
- server.ts is the canonical production entry with port fallback,
  Vercel support, Swagger, performance monitoring, and health checks

Phase 8.1.A: Architecture Normalization - Critical Fix
