+++ apps/backend/src/server.ts
@@ -69,9 +69,19 @@ app.use(
 app.use(compression() as unknown as RequestHandler);
 
 // CORS configuration
+const allowedOrigins = [
+  process.env.FRONTEND_URL || "http://localhost:3000",
+  "https://scholar-flow-ai.vercel.app",
+];
 app.use(
   cors({
-    origin: process.env.FRONTEND_URL || "http://localhost:3000",
+    origin: (origin, callback) => {
+      if (!origin || allowedOrigins.includes(origin)) {
+        callback(null, true);
+      } else {
+        callback(new Error("Not allowed by CORS"));
+      }
+    },
     credentials: true,
   }) as unknown as RequestHandler
 );
