+++ apps/backend/vercel.json
@@ -14,7 +14,7 @@
   "rewrites": [
     {
       "source": "/(.*)",
-      "destination": "/api/index.js"
+      "destination": "/api/server.js"
     }
   ]
 }
